Issue #84: extract interface from MathAstNode
