+++ src/main/java/parsevamath/tools/MathAstNode.java
@@ -28,153 +28,86 @@
 
 package parsevamath.tools;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 /**
  * The node type used in the homogeneous abstract syntax tree.
  */
-public class MathAstNode {
-
-    /**
-     * Token type of this MathAstNode.
-     */
-    private int tokenType;
-
-    /**
-     * Children of this MathAstNode.
-     */
-    private List<MathAstNode> children;
-
-    /**
-     * Position of this MathAstNode in expression.
-     */
-    private int columnNumber;
-
-    /**
-     * The parent MathAstNode of this MathAstNode.
-     */
-    private MathAstNode parent;
-
-    /**
-     * The text of this MathAstNode.
-     */
-    private String text;
-
-    /**
-     * Default constructor.
-     */
-    public MathAstNode() {
-        children = new ArrayList<>();
-    }
-
+public interface MathAstNode {
     /**
      * Get the parent node of this node.
      *
      * @return the parent of this node
      */
-    public final MathAstNode getParent() {
-        return parent;
-    }
+    MathAstNode getParent();
 
     /**
      * Set the parent node of this node.
      *
      * @param parent the MathAstNode to set
      */
-    public void setParent(MathAstNode parent) {
-        this.parent = parent;
-    }
+    void setParent(MathAstNode parent);
 
     /**
      * Adds one child to the child list of this MathAstNode.
      *
      * @param node the MathAstNode to add
      */
-    public void addChild(MathAstNode node) {
-        children.add(node);
-    }
+    void addChild(MathAstNode node);
 
     /**
      * Get the text of this node.
      *
      * @return string of text
      */
-    public String getText() {
-        return text;
-    }
+    String getText();
 
     /**
      * Set the text of this MathAstNode.
      *
      * @param text to be set
      */
-    public void setText(String text) {
-        this.text = text;
-    }
+    void setText(String text);
 
     /**
      * Gets type of this MathAstNode.
      *
      * @return type of node
      */
-    public int getTokenType() {
-        return tokenType;
-    }
+    int getTokenType();
 
     /**
      * Sets token type of this node.
      *
      * @param tokenType the type of token to set
      */
-    public void setTokenType(int tokenType) {
-        this.tokenType = tokenType;
-    }
+    void setTokenType(int tokenType);
 
     /**
      * Get an unmodifiable list of children from this MathAstNode.
      *
      * @return unmodifiable list
      */
-    public List<MathAstNode> getChildren() {
-        return Collections.unmodifiableList(children);
-    }
+    List<MathAstNode> getChildren();
 
     /**
      * Sets the children of this MathAstNode.
      *
      * @param children list of children
      */
-    public void setChildren(List<MathAstNode> children) {
-        this.children = Collections.unmodifiableList(children);
-    }
+    void setChildren(List<MathAstNode> children);
 
     /**
      * Gets position of this token.
      *
      * @return index of token in line
      */
-    public int getColumnNumber() {
-        return columnNumber;
-    }
+    int getColumnNumber();
 
     /**
      * Sets the column number of this token.
      *
      * @param columnNumber sets postion of token.
      */
-    public void setColumnNumber(int columnNumber) {
-        this.columnNumber = columnNumber;
-    }
-
-    @Override
-    public String toString() {
-        return "MathAstNode{" + "tokenType=" + tokenType
-            + ", children=" + children
-            + ", columnNumber=" + columnNumber
-            + ", parent=" + parent
-            + ", text='" + text + '\''
-            + '}';
-    }
+    void setColumnNumber(int columnNumber);
 }
